Move atlas/Constants to browser
